Convert standup plugin to read external config
All configuration was hardcoded for the first version. This
externalizes it to a yaml file that can be set in errbot's config.py,
under STANDUP_CONFIG_PATH, or it defaults to '/etc/standup/config.yaml'.

Also added a requirements.txt that will be auto-installed by errbot
when the plugin is loaded the first time.

Signed-off-by: Bradon Kanyid <bradon@kanyid.org>